Add files via upload
I put my Segmentation problem back in, I don't know what's wrong. Files Craft.rul, CraftWeapon.rul and Extrasprites.rul should be checked, the faulty machine name is Raptor, the faulty weapon name is Tachyon beam cannon. In new battle mode the test is fastest. There is also a minimal problem with the UFOpedia, it works, but for some reason it doesn't print the titles in the dissections, but the image displays fine, and the text that goes with it, it just doesn't print the titles.